fix other migrations

+++ database/migrations/2024_11_15_181540_remove_year_id_from_sales_target_versions.php
@@ -12,7 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('sales_target_versions', function (Blueprint $table) {
-            $table->dropColumn('year_id');
+            // Verificamos si la columna existe antes de intentar eliminarla
+            if (Schema::hasColumn('sales_target_versions', 'year_id')) {
+                $table->dropColumn('year_id');
+            }
         });
     }
 
@@ -22,7 +25,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('sales_target_versions', function (Blueprint $table) {
-            //
+            // Si necesitas recrear la columna en el rollback
+            if (!Schema::hasColumn('sales_target_versions', 'year_id')) {
+                $table->unsignedBigInteger('year_id')->after('id');
+            }
         });
     }
 };
